refactor: rename app-boot.ts to appBoot.ts and update import paths
