+++ packages/webapp/app/layout.tsx
@@ -1,8 +1,8 @@
 import type { PropsWithChildren, ReactElement } from 'react';
 import React from 'react';
 import Head from 'next/head';
+import { AppHeadMetas } from '@dailydotdev/shared/src/features/common/components/AppHeadMetas';
 import Providers from './providers';
-import { AppHeadMetas } from '../../shared/src/features/common/components/AppHeadMetas';
 import '@dailydotdev/shared/src/styles/globals.css';
 
 // todo: Improve this by moving metas to metadta object as suggested in https://nextjs.org/docs/app/building-your-application/optimizing/metadata
